+++ docs_manual/conf.py
@@ -108,7 +108,7 @@ html_theme = "sphinx_rtd_theme"
 # These paths are either relative to html_static_path or fully qualified paths
 # (eg. https://...)
 html_css_files = [
-    'css/custom.css',
+    "css/custom.css",
 ]
 
 # Theme options are theme-specific and customize the look and feel of a theme