+++ HISTORY.rst
@@ -44,6 +44,8 @@ End-User Summary
 - Added option to toggle displaying of logs during filtration, by default they are hidden.
 - Fixed broken displaying of inhouse frequencies in variant detail view.
 - Added variant annotation list (comments, flags, ACMG ratings) to project-wide info page.
+- Row in filter results now turns gray when any flag is set (except bookmark flag; summary flag still colours in other colour).
+- Fixed bug where comments and flags in variant details weren't updated when the variant details have been opened before.
 
 Full Change List
 ================
@@ -88,6 +90,8 @@ Full Change List
 - Added option to toggle displaying of logs during filtration, by default they are hidden.
 - Fixed broken displaying of inhouse frequencies in variant detail view.
 - Added variant annotation list (comments, flags, ACMG ratings) to project-wide info page.
+- Row in filter results now turns gray when any flag is set (except bookmark flag; summary flag still colours in other colour).
+- Fixed bug where comments and flags in variant details weren't updated when the variant details have been opened before.
 
 -------
 v0.20.0